+++ EssentialFeed/EssentialFeed/Feed API/HTTPClient.swift
@@ -1,0 +1,10 @@
+import Foundation
+
+public enum HTTPClientResult {
+    case success(Data, HTTPURLResponse)
+    case failure(Error)
+}
+
+public protocol HTTPClient {
+    func get(from url: URL, completion: @escaping (HTTPClientResult) -> Void)
+}
