+++ EssentialFeed/EssentialFeed/Feed API/RemoteFeedLoader.swift
@@ -1,14 +1,5 @@
 import Foundation
 
-public enum HTTPClientResult {
-    case success(Data, HTTPURLResponse)
-    case failure(Error)
-}
-
-public protocol HTTPClient {
-    func get(from url: URL, completion: @escaping (HTTPClientResult) -> Void)
-}
-
 public class RemoteFeedLoader {
     private let client: HTTPClient
     private let url: URL
