+++ EssentialFeed/EssentialFeedTests/RemoteFeedLoaderTests.swift
@@ -155,4 +155,3 @@ final class RemoteFeedLoaderTests: XCTestCase {
     }
     
 }
-
